Added support for the ticket scanner functionality. This uses the barcode library to scan the QR code and determine whether the details are correct. This checks whether the details match and whether or not the ticket matches the venue/gig. If it doesn't the ticket is not deleted. Other checks are included. Also fixed the ticket purchase dialog width. Now need to add ticket cost to the gig creation/edit screens so they can then be pulled from the database at the point of purchase.

+++ app/src/main/java/com/liamd/giggity_app/FanUserGigDetailsFragment.java
@@ -272,7 +272,7 @@ public class FanUserGigDetailsFragment extends Fragment implements OnMapReadyCal
         ticketPickerDialog.setTitle("Ticket Picker");
         ticketPickerDialog.setContentView(R.layout.ticket_purchase_dialog_layout);
 
-        final double ticketCost = 10;
+        final int ticketCost = 10;
 
         // Initialise dialog visual components
         TextView mGigNameTextView = (TextView) ticketPickerDialog.findViewById(R.id.gigNameTextView);
